Fix Calc  Adding Double commit

+++ Cacl/Calcu.cs
@@ -8,19 +8,19 @@ namespace Cacl
 {
     public class Calcu
     {
-        public static double Sum(int a, int b)
+        public static double Sum(double a, double b)
         {
             return a + b;
         }
-        public static double Min(int a, int b)
+        public static double Min(double a, double b)
         {
             return a - b;
         }
-        public static double Div(int a, int b)
+        public static double Div(double a, double b)
         {
             return a / b;
         }
-        public static double Mul(int a, int b)
+        public static double Mul(double a, double b)
         {
             return a * b;
         }
